SW-3951 Refactor services
